Fix SIGSEGV on -Ox by properly handling PHINodes
Properly handle blocks that contain PHINodes by translating them to
arguments for the extracted BasicBlock. Once replaced by an argument,
a PHINode can be removed.

This commit also ensures the entry block ends with an unconditional
branch in order to have a single root node for the postorder traversal
in the first pass.